backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
sqlalchemy==2.0.43
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic[email]==2.9.2
pydantic-settings==2.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.2
httpx==0.28.1
pyotp==2.9.0
cryptography==45.0.7
google-api-python-client==2.147.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
twilio==9.3.1
aiosmtplib==3.0.2
